biosignals/requirements.txt
AidlabSDK==1.3.26
astropy==5.2.2
bidict==0.22.1
biosppy==1.0.0
bleak==0.14.3
contourpy==1.0.7
cycler==0.11.0
dbus-next==0.2.3
fonttools==4.39.4
future==0.18.3
h5py==3.8.0
hrv-analysis==1.0.4
importlib-resources==5.12.0
joblib==1.2.0
kiwisolver==1.4.4
matplotlib==3.7.1
nolds==0.5.2
numpy==1.24.3
opencv-python==4.7.0.72
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
pyerfa==2.0.0.3
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
scikit-learn==1.2.2
scipy==1.10.1
shortuuid==1.0.11
six==1.16.0
threadpoolctl==3.1.0
tzdata==2023.3
zipp==3.15.0
